+++ scripts/nmea_socket_driver.py
@@ -37,7 +37,7 @@ import sys
 
 import rclpy
 
-from nmea_navsat_driver.driver import Ros2NMEADriver
+from src.libnmea_navsat_driver.driver import Ros2NMEADriver
 
 
 def main(args=None):
