+++ scripts/nmea_topic_serial_reader.py
@@ -36,9 +36,8 @@ import serial
 
 import rclpy
 
-import time
 from nmea_msgs.msg import Sentence
-from nmea_navsat_driver.driver import Ros2NMEADriver
+from src.libnmea_navsat_driver.driver import Ros2NMEADriver
 
 
 def main(args=None):
