+++ src/libnmea_navsat_driver/checksum_utils.py
