+++ src/libnmea_navsat_driver/driver.py
@@ -38,8 +38,8 @@ from rclpy.node import Node
 from sensor_msgs.msg import NavSatFix, NavSatStatus, TimeReference
 from geometry_msgs.msg import TwistStamped, QuaternionStamped
 from transforms3d.euler import euler2quat as quaternion_from_euler
-from nmea_navsat_driver.checksum_utils import check_nmea_checksum
-from nmea_navsat_driver import parser
+from src.libnmea_navsat_driver.checksum_utils import check_nmea_checksum
+from src.libnmea_navsat_driver import parser
 
 
 class Ros2NMEADriver(Node):
