+++ src/libnmea_navsat_driver/parser.py
